pyproject.toml
[tool.poetry]
name = "mindmeet-padb"
version = "0.1.0"
description = ""
authors = ["Jackson Serafim <jacksonvieiira12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg = "^3.1.12"
pytest = "^7.4.2"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
python-decouple = "^3.8"
ruff = "^0.1.1"
psycopg-binary = "^3.1.12"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
